+++ cypress/locators/action-popover/index.js
@@ -2,6 +2,9 @@ import {
   ACTION_POPOVER_BUTTON,
   ACTION_POPOVER_DATA_COMPONENT,
   ACTION_POPOVER_SUBMENU,
+  ACTION_POPOVER_MENU_ITEM_ICON,
+  ACTION_POPOVER_MENU_ITEM_INNER_TEXT,
+  ACTION_POPOVER_MENU_ITEM_CHEVRON,
   ACTION_POPOVER_WRAPPER,
 } from "./locators";
 
@@ -18,6 +21,12 @@ export const actionPopoverInnerItem = (index) =>
     .first();
 export const actionPopoverSubmenu = (index) =>
   cy.get(ACTION_POPOVER_SUBMENU).eq(1).children().eq(index).find("button");
+export const actionPopoverMenuItemIcon = () =>
+  cy.get(ACTION_POPOVER_MENU_ITEM_ICON);
+export const actionPopoverMenuItemInnerText = () =>
+  cy.get(ACTION_POPOVER_MENU_ITEM_INNER_TEXT);
+export const actionPopoverMenuItemChevron = () =>
+  cy.get(ACTION_POPOVER_MENU_ITEM_CHEVRON);
 export const actionPopoverSubmenuByIndex = () =>
   cy.get(ACTION_POPOVER_SUBMENU).eq(1);
 export const actionPopoverWrapper = () => cy.get(ACTION_POPOVER_WRAPPER);